Updates upload scripts to have clearer names and documentation.
