+++ api-service/api-server/src/setup/custom_uploaders/upload_toxicity_subtypes.ts
@@ -17,38 +17,42 @@ limitations under the License.
 /*
 Usage:
 
-node build/server/setup/upload_toanswer.js \
-  --file="./tmp/real_job/toanswer_mini_x10.json" \
-  --question_group="wp_v1_x10k" \
+node build/server/setup/upload_with_answers.js \
+  --file="./tmp/real_job/with_answers_mini_10x.json" \
+  --question_group="wp_x10k_test" \
+  --training_fraction=0.2 \
   --action=print
 
-node build/server/setup/upload_toanswer.js \
-  --file="./tmp/real_job/toanswer.json" \
-  --question_group="wp_v1_x10k" \
+node build/server/setup/upload_with_answers.js \
+  --file="./tmp/real_job/with_answers.json" \
+  --question_group="wp_x10k_test" \
+  --training_fraction=0.2 \
   --action=insert
 
-node build/server/setup/upload_toanswer.js \
-  --file="./tmp/real_job/toanswer.json" \
+node build/server/setup/upload_with_answers.js \
+  --file="./tmp/real_job/with_answers.json" \
   --question_group="wp_v1_x10k" \
-  --action=print \
-  --from_line 2600 --to_line=2610
+  --training_fraction=0.2 \
+  --action=update
 
-ts-node src/setup/upload_toanswer.ts \
-  --file="./tmp/foo.json" \
-  --spanner_db="ptoxic" \
-  --question_group="erica-demo-requests" \
-  --action=print
+node build/server/setup/upload_with_answers.js \
+  --file="./tmp/real_job/with_answers.json" \
+  --question_group="wp_x10k_test" \
+  --training_fraction=0.2 \
+  --action=print \
+  --from_line 2
 */
 
 import * as spanner from '@google-cloud/spanner';
 import * as fs from 'fs';
+import * as readline from 'readline';
+import * as seedrandom from 'seedrandom';
 import * as yargs from 'yargs';
 
-import * as crowdsourcedb from '../cs_db';
-// import * as readline from 'readline';
-import * as db_types from '../db_types';
+import * as crowdsourcedb from '../../cs_db';
+import * as db_types from '../../db_types';
 
-import * as util from './util';
+import * as util from '../util';
 
 // TODO(ldixon): fix import of this.
 const hashjs = require('hash.js');
@@ -58,12 +62,31 @@ interface Params {
   file: string, question_group: string, gcloud_project_id: string,
       spanner_instance: string, spanner_db: string,
       action: 'print'|'update'|'insert'|'count', random_seed: string;
+  training_fraction: number;  // float.
   from_line: number|null;
   to_line: number|null;
 }
+;
 
+// type DataShape = wpconvlib.Conversation;
 interface DataShape {
-  id: string, text: string,
+  frac_neg: string
+  obscene: string;
+  threat: string;
+  insult: string;
+  identity_hate: string;
+  rev_id: string;
+  comment_text: string;
+}
+;
+
+//
+function makeScoreEnum(frac: number) {
+  return {
+    notatall: (frac < 0.3) ? 1 : (frac < 0.5 ? 0 : -1),
+        somewhat: (frac < 0.2) ? -0.5 : (frac < 0.8 ? 1 : -0.5),
+        very: (frac < 0.5) ? -1 : (frac < 0.7 ? 0 : 1),
+  }
 }
 
 class ArgError extends Error {}
@@ -71,6 +94,10 @@ class ArgError extends Error {}
 function checkArguments(args: Params): void {
   if (!fs.existsSync(args.file)) {
     throw new ArgError('--file argument does not exist: ' + args.file);
+  }
+  if (typeof (args.training_fraction) !== 'number') {
+    throw new ArgError(
+        '--training_fraction must be a float, not: ' + args.training_fraction);
   }
   if (!db_types.valid_id_regexp.test(args.question_group)) {
     throw new ArgError(
@@ -87,20 +114,51 @@ function checkArguments(args: Params): void {
   }
 }
 
-function questionRowFromDatum(datum: DataShape): db_types.QuestionRow {
-  let id = hashjs.sha256().update(datum.id + ':' + datum.text).digest('hex');
-  let question: db_types.Question = {id: id, text: datum.text} as any;
+function questionRowFromDatum(
+    datum: DataShape,
+    question_type: 'training'|'test'|'toanswer'): db_types.QuestionRow {
+  let id = hashjs.sha256()
+               .update(datum.rev_id + ':' + datum.comment_text)
+               .digest('hex');
+  let question: db_types.Question = {id: id, text: datum.comment_text} as any;
+
+  let toxicity_scores = {enum: makeScoreEnum(parseFloat(datum.frac_neg))};
+  let readableAndInEnglishScores = {optional: true, enum: {yes: 1, no: -1}};
+  let obscene_scores = {
+    optional: true,
+    enum: makeScoreEnum(parseFloat(datum.obscene))
+  };
+  let hate_scores = {
+    optional: true,
+    enum: makeScoreEnum(parseFloat(datum.obscene))
+  };
+  let threat_scores = {
+    optional: true,
+    enum: makeScoreEnum(parseFloat(datum.obscene))
+  };
+  let insult_scores = {
+    optional: true,
+    enum: makeScoreEnum(parseFloat(datum.obscene))
+  };
 
   let questionRow: db_types.QuestionRow = {
-    accepted_answers: null,
+    accepted_answers: {
+      readableAndInEnglish: readableAndInEnglishScores,
+      toxic: toxicity_scores,
+      obscene: obscene_scores,
+      identityHate: hate_scores,
+      threat: threat_scores,
+      insult: insult_scores,
+      comments: {optional: true, freeStringConstScore: 0},
+    },
     question: question,
     question_group_id: args.question_group,
     question_id: id,
-    type: 'toanswer',
+    // TODO(ldixon): make nice typescript interence helper function.
+    type: question_type,
   };
   return questionRow;
 }
-
 
 async function main(args: Params) {
   checkArguments(args);
@@ -110,31 +168,41 @@ async function main(args: Params) {
   let spannerDatabase =
       spannerInstance.database(args.spanner_db, {keepAlive: 5});
   let db = new crowdsourcedb.CrowdsourceDB(spannerDatabase);
+  let random = seedrandom(args.random_seed);
+
+  // let fileAsString = fs.readFileSync(args.file, 'utf8');
 
   let total_added = 0;
+  let training_added = 0;
+  let test_added = 0;
   let batcher = new util.Batcher<db_types.QuestionRow>(
       async (batch: db_types.QuestionRow[]) => {
         if (args.action === 'update') {
           await db.updateQuestions(batch);
         } else if (args.action === 'insert') {
-          await db.addQuestions(batch).catch((e) => {
-            console.error('Failed at batch: ' + JSON.stringify(batch, null, 2));
-            throw e;
-          });
+          await db.addQuestions(batch);
         } else if (args.action === 'print') {
-          console.log(JSON.stringify(batch, null, 2));
+          console.log(JSON.stringify(batch));
         } else if (args.action === 'count') {
         } else {
           throw new Error('invalid action: ' + args.action);
         }
-        console.log(`looked at question batch size: ${batch.length}`);
+        console.log(`questions in batch: ${batch.length}`);
         console.log(`total questions is now: ${total_added}`);
         total_added += batch.length;
       },
       10);
 
   await util.applyToLinesOfFile(args.file, async (line: string) => {
-    let question = questionRowFromDatum(JSON.parse(line));
+    let question_type: 'test'|'training';
+    if (random() >= args.training_fraction) {
+      question_type = 'test';
+      test_added += 1;
+    } else {
+      question_type = 'training';
+      training_added += 1;
+    }
+    let question = questionRowFromDatum(JSON.parse(line), question_type);
     return batcher.add(question);
   }, {from_line: args.from_line, to_line: args.to_line});
 
@@ -143,6 +211,8 @@ async function main(args: Params) {
   await db.close();
 
   console.log(`Completed. Added questions: ${total_added}`);
+  console.log(`Training: ${training_added}`);
+  console.log(`Test: ${test_added}`);
 }
 
 let args =
@@ -151,6 +221,9 @@ let args =
           describe:
               'A random seed to be used to select fraction of comments to be used for training'
         })
+        .option(
+            'training_fraction',
+            {describe: 'Fraction of examples given to be for training'})
         .option(
             'question_type',
             {describe: 'Question type, one of: training | test | toanswer'})
@@ -173,9 +246,10 @@ let args =
         .default('action', 'print')
         .default('gcloud_project_id', 'wikidetox')
         .default('spanner_instance', 'crowdsource')
+        .default('spanner_db', 'testdb')
         .demandOption(
-            ['file', 'question_group', 'spanner_db'],
-            'Please provide at least --file, --question_group, and --spanner_db.')
+            ['file', 'question_group', 'training_fraction'],
+            'Please provide at least --file and --question_group, and --training_fraction.')
         .help()
         .argv;
